+++ R/utils.R
@@ -8,6 +8,7 @@ select_loc <- function(x,
                        allow_rename = TRUE,
                        allow_empty = TRUE,
                        allow_predicates = TRUE,
+                       error_arg = NULL,
                        error_call = current_env()) {
   check_dots_empty()
 
@@ -21,6 +22,7 @@ select_loc <- function(x,
     allow_rename = allow_rename,
     allow_empty = allow_empty,
     allow_predicates = allow_predicates,
+    error_arg = error_arg,
     error_call = error_call
   )
 }
@@ -51,6 +53,7 @@ relocate_loc <- function(x,
                          name_spec = NULL,
                          allow_rename = TRUE,
                          allow_empty = TRUE,
+                         error_arg = NULL,
                          before_arg = "before",
                          after_arg = "after",
                          error_call = current_env()) {
@@ -65,6 +68,7 @@ relocate_loc <- function(x,
     name_spec = name_spec,
     allow_rename = allow_rename,
     allow_empty = allow_empty,
+    error_arg = error_arg,
     before_arg = before_arg,
     after_arg = after_arg,
     error_call = error_call