+++ tests/testthat/test-eval-relocate.R
@@ -177,11 +177,22 @@ test_that("can forbid empty selections", {
   x <- c(a = 1, b = 2, c = 3)
 
   expect_snapshot({
-    (expect_error(relocate_loc(x, allow_empty = FALSE)))
+    (expect_error(relocate_loc(x, allow_empty = FALSE, error_arg = "...")))
     (expect_error(relocate_loc(mtcars, integer(), allow_empty = FALSE)))
     (expect_error(relocate_loc(mtcars, starts_with("z"), allow_empty = FALSE)))
   })
 })
+
+test_that("can forbid empty selections", {
+  x <- c(a = 1, b = 2, c = 3)
+  
+  expect_snapshot(
+    error = TRUE, {
+    relocate_loc(mtcars, before = integer(), allow_empty = FALSE)
+    relocate_loc(mtcars, starts_with("z"), allow_empty = FALSE)
+  }, cnd_class = TRUE)
+})
+
 
 test_that("`before` and `after` forbid renaming", {
   x <- c(a = 1, b = 2, c = 3)